backend/files/urls.py
from django.urls import path, include
from rest_framework.routers import DefaultRouter
from .views import FileViewSet

router = DefaultRouter()
router.register(r'files', FileViewSet, basename='files')

urlpatterns = [
    path('', include(router.urls)),
] 